Selects: remove old components
