--- src/components/forms/select-headless/multi/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
