--- src/components/forms/select-headless/single/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
