--- src/components/forms/select/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
