--- src/components/forms/select/menu/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
