+++ src/components/forms/select/multi/component.tsx
@@ -1,310 +1,221 @@
-import { useEffect, useRef, useMemo, FC } from 'react';
-
-import { createPortal } from 'react-dom';
-import { usePopper } from 'react-popper';
+import React, { FC, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import cx from 'classnames';
 
-// Downshift;
-import { useSelect, useMultipleSelection } from 'downshift';
+import { Listbox, Transition } from '@headlessui/react';
 
 import Checkbox from 'components/forms/checkbox';
-// Popper
-import {
-  flipModifier,
-  hideModifier,
-  sameWidthModifier,
-  offsetModifier,
-} from 'components/forms/select/constants/popper-modifiers';
 import THEME from 'components/forms/select/constants/theme';
-import Menu from 'components/forms/select/menu';
-import Toggle from 'components/forms/select/toggle';
-import { SelectProps, SelectOptionProps } from 'components/forms/select/types';
+import Icon from 'components/icon';
+import Loading from 'components/loading';
 
-export const MultiSelect: FC<SelectProps> = ({
-  theme = 'dark',
-  size = 'base',
-  maxHeight = 300,
-  status,
-  prefix,
-  options = [],
-  values,
-  initialValues = [],
-  disabled = false,
-  multiple = true,
-  placeholder,
-  clearSelectionActive = true,
-  clearSelectionLabel = 'Clear selection',
-  batchSelectionActive,
-  batchSelectionLabel = 'Select all',
-  onSelect,
-  onFocus,
-  onBlur,
-}: SelectProps) => {
-  const triggerRef = useRef();
-  const menuRef = useRef();
+import CHEVRON_DOWN_SVG from 'svgs/ui/arrow-down.svg?sprite';
+import CHEVRON_UP_SVG from 'svgs/ui/arrow-up.svg?sprite';
 
-  const getOptions: SelectOptionProps[] = useMemo(
-    () => [
-      ...(clearSelectionActive
-        ? [
-            {
-              value: 'batch-clear-selection',
-              label: clearSelectionLabel,
-              enabled: false,
-              checkbox: false,
-            },
-          ]
-        : []),
-      ...(batchSelectionActive
-        ? [
-            {
-              value: 'batch-selection',
-              label: batchSelectionLabel,
-              enabled: false,
-              checkbox: false,
-            },
-          ]
-        : []),
-      ...options.map((o) => ({ ...o, checkbox: true, enabled: true })),
-    ],
-    [options, clearSelectionActive, clearSelectionLabel, batchSelectionActive, batchSelectionLabel]
+import type { MultiSelectProps } from './types';
+
+export const Select: FC<MultiSelectProps> = (props: MultiSelectProps) => {
+  const {
+    batchSelectionActive = false,
+    batchSelectionLabel = 'Select all',
+    clearSelectionActive = false,
+    clearSelectionLabel = 'Clear all',
+    disabled = false,
+    options,
+    placeholder = 'Select...',
+    loading,
+    size = 'base',
+    theme = 'dark',
+    state = 'none',
+    values,
+    onChange,
+  } = props;
+  const ref = useRef(null);
+  const initialValues = values || [];
+  const [selected, setSelected] = useState(initialValues);
+
+  const SELECTED = useMemo(() => {
+    if (loading) return 'Loading...';
+
+    if (!selected.length) return placeholder || 'Select items';
+
+    // if (selected.length === options.length) return placeholder;
+
+    if (selected.length === 1) {
+      const option = options.find((o) => o.value === selected[0]);
+      if (option) return option.label;
+      return null;
+    }
+
+    if (selected.length === options.length) return `All items selected`;
+
+    if (selected.length > 1) return `Selected items (${selected.length})`;
+
+    return null;
+  }, [loading, options, placeholder, selected]);
+
+  useEffect(() => {
+    if (values) {
+      setSelected(values);
+    }
+  }, [values]);
+
+  const handleSelect = useCallback(
+    (v) => {
+      setSelected(v);
+      if (onChange) {
+        onChange(v);
+      }
+    },
+    [onChange]
   );
 
-  const getOptionsEnabled = useMemo(() => {
-    const opts = getOptions.filter((op) => !op.disabled && op.enabled);
-    return opts;
-  }, [getOptions]);
-
-  const getInitialSelected = getOptions.filter((o) => initialValues.includes(`${o.value}`));
-
-  const getSelected = values ? getOptions.filter((o) => values.includes(`${o.value}`)) : null;
-
-  const isSelected = (selected: SelectOptionProps, selectedItms: SelectOptionProps[]) =>
-    selectedItms.some((i) => i.value === selected.value);
-
-  const handleSelectedItem = ({
-    option,
-    selectedItems,
-    addSelectedItem,
-    removeSelectedItem,
-    setSelectedItems,
-    reset,
-  }) => {
-    switch (option.value) {
-      case 'batch-clear-selection':
-        reset();
-        break;
-      case 'batch-selection':
-        setSelectedItems(getOptionsEnabled);
-        break;
-      default:
-        if (option.disabled) {
-          break;
-        }
-
-        if (isSelected(option, selectedItems)) {
-          removeSelectedItem(option);
-          break;
-        } else {
-          addSelectedItem(option);
-          break;
-        }
+  const handleSelectAll = useCallback(() => {
+    const allOptions = options.map((o) => o.value);
+    setSelected(allOptions);
+    if (onChange) {
+      onChange(allOptions);
     }
-  };
+  }, [onChange, options]);
 
-  const {
-    getDropdownProps,
-    addSelectedItem,
-    removeSelectedItem,
-    setSelectedItems,
-    selectedItems,
-    reset,
-  } = useMultipleSelection<SelectOptionProps>({
-    ...(!!getSelected && {
-      selectedItems: getSelected,
-    }),
-    ...(!!getInitialSelected && {
-      initialSelectedItems: getInitialSelected,
-    }),
-    itemToString: (item) => item.label, // How the selected options is announced to screen readers
-    stateReducer: (st, actionAndChanges) => {
-      const { changes, type } = actionAndChanges;
-      if (
-        type === useMultipleSelection.stateChangeTypes.FunctionAddSelectedItem ||
-        type === useMultipleSelection.stateChangeTypes.FunctionRemoveSelectedItem ||
-        type === useMultipleSelection.stateChangeTypes.FunctionSetSelectedItems ||
-        type === useMultipleSelection.stateChangeTypes.FunctionReset
-      ) {
-        onSelect(changes.selectedItems);
-      }
-
-      return changes;
-    },
-  });
-
-  const { isOpen, highlightedIndex, getToggleButtonProps, getMenuProps, getItemProps, closeMenu } =
-    useSelect<SelectOptionProps>({
-      items: getOptions,
-      itemToString: (item) => item.label, // How the selected options is announced to screen readers
-      stateReducer: (st, actionAndChanges) => {
-        const { type, changes } = actionAndChanges;
-        const { selectedItem: option } = changes;
-
-        switch (type) {
-          case useSelect.stateChangeTypes.MenuKeyDownEnter:
-          case useSelect.stateChangeTypes.ItemClick:
-            handleSelectedItem({
-              option,
-              selectedItems,
-              addSelectedItem,
-              removeSelectedItem,
-              setSelectedItems,
-              reset,
-            });
-
-            return {
-              ...changes,
-              highlightedIndex: st.highlightedIndex,
-              isOpen: true,
-            };
-          default:
-            return changes;
-        }
-      },
-    });
-
-  // 'usePopper'
-  const { styles, attributes, update } = usePopper(triggerRef.current, menuRef.current, {
-    placement: 'bottom',
-    // strategy: 'fixed',
-    modifiers: [offsetModifier, flipModifier, hideModifier, sameWidthModifier],
-  });
-
-  // Hide menu if reference is outside the boundaries
-  const referenceHidden =
-    attributes?.popper?.['data-popper-reference-hidden'] ||
-    attributes?.popper?.['data-popper-reference-scaped'];
-  useEffect(() => {
-    if (referenceHidden) {
-      closeMenu();
+  const handleClearAll = useCallback(() => {
+    setSelected([]);
+    if (onChange) {
+      onChange([]);
     }
-  }, [referenceHidden, closeMenu]);
-
-  useEffect(() => {
-    if (update) update();
-  }, [isOpen, update]);
+  }, [onChange]);
 
   return (
     <div
       className={cx({
-        'w-full leading-tight overflow-hidden': true,
+        'w-full': true,
         [THEME[theme].container]: true,
-        [THEME[theme].closed]: true,
-        [THEME.states[status]]: true,
       })}
     >
-      <div className="relative w-full" ref={triggerRef}>
-        <Toggle
-          options={getOptionsEnabled}
-          theme={theme}
-          size={size}
-          status={status}
-          prefix={prefix}
-          disabled={disabled}
-          multiple
-          opened={isOpen}
-          selectedItems={selectedItems}
-          placeholder={placeholder}
-          getToggleButtonProps={getToggleButtonProps}
-          getDropdownProps={getDropdownProps}
-        />
-      </div>
-
-      {/* Menu */}
-      {createPortal(
-        <div
-          className={cx({
-            'z-50': true,
-            // The content of `<Menu />` must always be in the DOM so that Downshift can get the ref
-            // to the `<ul />` element through `getMenuProps`
-            invisible: !isOpen,
-          })}
-          ref={menuRef}
-          style={styles.popper}
-          {...attributes.popper}
-        >
-          <Menu
-            theme={theme}
-            size={size}
-            status={status}
-            disabled={disabled}
-            multiple
-            opened={isOpen}
-            attributes={attributes}
-          >
-            <Toggle
-              options={getOptionsEnabled}
-              theme={theme}
-              size={size}
-              status={status}
-              prefix={prefix}
-              disabled={disabled}
-              multiple={multiple}
-              opened={isOpen}
-              selectedItems={selectedItems}
-              placeholder={placeholder}
-              getToggleButtonProps={getToggleButtonProps}
-              getDropdownProps={getDropdownProps}
-            />
-
-            <ul
-              {...getMenuProps({ onFocus, onBlur })}
-              className={cx({
-                'py-1 overflow-y-auto overflow-x-hidden': true,
-              })}
-              style={{
-                maxHeight,
-              }}
-            >
-              {getOptions.map((option, index) => (
-                <li
+      <Listbox
+        as="div"
+        className="space-y-1"
+        disabled={disabled}
+        value={selected}
+        multiple
+        onChange={handleSelect}
+      >
+        {({ open }) => (
+          <>
+            <div className="relative space-y-3" ref={ref}>
+              <span className="inline-block w-full">
+                <Listbox.Button
                   className={cx({
-                    'px-4 py-1 mt-0.5 cursor-pointer relative': true,
-                    [THEME[theme].item.base]: highlightedIndex !== index,
-                    [THEME[theme].item.disabled]: option.disabled,
-                    [THEME[theme].item.highlighted]:
-                      (highlightedIndex === index && !option.disabled) ||
-                      isSelected(option, selectedItems),
+                    [THEME[theme].button.base]: true,
+                    [THEME[theme].button.states.disabled]: disabled,
+                    [THEME[theme].button.states.valid]: state === 'valid',
+                    [THEME[theme].button.states.error]: state === 'error',
+                    [THEME.sizes[size]]: true,
                   })}
-                  key={`${option.value}`}
-                  {...getItemProps({ item: option, index, disabled: option.disabled })}
                 >
-                  <span
-                    className={cx({
-                      'ml-6': !!option.checkbox,
-                    })}
-                  >
-                    {option.label}
-                  </span>
-
-                  {option.checkbox && (
-                    <Checkbox
-                      className="absolute bg-opacity-0 left-4 top-1.5"
-                      checked={isSelected(option, selectedItems)}
-                      disabled={option.disabled}
-                      onChange={() => null}
+                  <span className="block truncate">{SELECTED}</span>
+                  <span className="absolute inset-y-0 right-0 flex items-center pr-3 pointer-events-none">
+                    <Loading
+                      visible={loading}
+                      className={THEME[theme].loading}
+                      iconClassName="w-3 h-3"
                     />
-                  )}
-                </li>
-              ))}
-            </ul>
-          </Menu>
-        </div>,
-        document.body
-      )}
+
+                    {!loading && (
+                      <Icon
+                        icon={open ? CHEVRON_UP_SVG : CHEVRON_DOWN_SVG}
+                        className={cx({
+                          'w-3 h-3': true,
+                        })}
+                      />
+                    )}
+                  </span>
+                </Listbox.Button>
+              </span>
+
+              <Transition
+                unmount={false}
+                show={open}
+                leave="transition ease-in duration-100"
+                leaveFrom="opacity-100"
+                leaveTo="opacity-0"
+                className={cx({
+                  'z-10 absolute w-full overflow-y-auto rounded-lg shadow-lg min-w-[250px]': true,
+                })}
+              >
+                <Listbox.Options
+                  static
+                  className={cx({
+                    'overflow-y-auto text-base leading-6 max-h-60 focus:outline-none': true,
+                    [THEME[theme].menu]: true,
+                  })}
+                >
+                  <div className="flex px-5 space-x-5 text-sm">
+                    {batchSelectionActive && (
+                      <button
+                        className="py-2 text-left underline text-grey-20 whitespace-nowrap"
+                        type="button"
+                        onClick={handleSelectAll}
+                      >
+                        {batchSelectionLabel}
+                      </button>
+                    )}
+
+                    {clearSelectionActive && (
+                      <button
+                        className="py-2 text-left underline whitespace-nowrap"
+                        type="button"
+                        onClick={handleClearAll}
+                      >
+                        {selected.length < 1 && clearSelectionLabel}
+                        {selected.length >= 1 &&
+                          selected.length !== options.length &&
+                          `${clearSelectionLabel} (${selected.length} Selected)`}
+                        {selected.length === options.length &&
+                          `${clearSelectionLabel} (All selected)`}
+                      </button>
+                    )}
+                  </div>
+
+                  {options.map((opt) => {
+                    return (
+                      <Listbox.Option key={opt.value} value={opt.value}>
+                        {({ active: a, selected: s, disabled: d }) => (
+                          <div
+                            className={cx({
+                              'flex items-center space-x-2 cursor-pointer select-none relative py-2 pl-5 pr-4':
+                                true,
+                              [THEME[theme].item.base]: true,
+                              [THEME[theme].item.active]: a,
+                              [THEME[theme].item.disabled]: d,
+                            })}
+                          >
+                            <Checkbox
+                              className="cursor-pointer focus:text-black focus:ring-black checked:bg-black"
+                              checked={s}
+                              readOnly
+                            />
+
+                            <span
+                              className={cx({
+                                'font-semibold block line-clamp-2': true,
+                              })}
+                            >
+                              {opt.label}
+                            </span>
+                          </div>
+                        )}
+                      </Listbox.Option>
+                    );
+                  })}
+                </Listbox.Options>
+              </Transition>
+            </div>
+          </>
+        )}
+      </Listbox>
     </div>
   );
 };
 
-export default MultiSelect;
+export default Select;
