--- src/components/forms/select/toggle/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
